covid19datapool_dbsync/requirements.txt
Flask==1.1.1
gitpython
google-cloud-firestore
google-cloud-tasks==1.3.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
openpyxl
python-dateutil
pycountry
psycopg2-binary
